Created initial Controllers and Types.

+++ src/Entity/Product.php
@@ -33,7 +33,7 @@ class Product
 
     #[ORM\Column(
         nullable: true)]
-    private ?int $rating = null;
+    private ?float $rating = null;
 
     public function __construct()
     {
@@ -99,12 +99,12 @@ class Product
         return $this;
     }
 
-    public function getRating(): ?int
+    public function getRating(): ?float
     {
         return $this->rating;
     }
 
-    public function setRating(?int $rating): self
+    public function setRating(?float $rating): self
     {
         $this->rating = $rating;
